Rename types files
