+++ packages/graphql_server/src/api/openAIApi.ts
@@ -1,4 +1,4 @@
-import { OpenAIResponse } from '../../types/openAIApi.types'
+import { OpenAIResponse } from '../../types/openAIApi'
 
 /** Creates the ELI5 of the repository by passing the readme to chatgpt
  * @param {string} readME readMe file of the repository as a string
