+++ packages/graphql_server/types/githubScraping.d.ts
