+++ packages/graphql_server/types/starHistory.d.ts
