+++ packages/graphql_server/types/supabase.d.ts
